Move JPA types to separate module (#500)
`polaris-core` should be rather generic and not depend on JPA per se, so moving the types to a new module + package.
