+++ extension/persistence/jpa-model/src/main/java/org/apache/polaris/jpa/models/ModelSequenceId.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.polaris.core.persistence.models;
+package org.apache.polaris.jpa.models;
 
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
