+++ gradle/projects.main.properties
@@ -21,4 +21,5 @@
 polaris-core=polaris-core
 polaris-service=polaris-service
 polaris-eclipselink=extension/persistence/eclipselink
+polaris-jpa-model=extension/persistence/jpa-model
 aggregated-license-report=aggregated-license-report
