+++ helm/polaris/ci/fixtures/persistence.yaml
@@ -29,13 +29,13 @@ stringData:
       xsi:schemaLocation="http://java.sun.com/xml/ns/persistence http://java.sun.com/xml/ns/persistence/persistence_2_0.xsd">
       <persistence-unit name="polaris" transaction-type="RESOURCE_LOCAL">
         <provider>org.eclipse.persistence.jpa.PersistenceProvider</provider>
-        <class>org.apache.polaris.core.persistence.models.ModelEntity</class>
-        <class>org.apache.polaris.core.persistence.models.ModelEntityActive</class>
-        <class>org.apache.polaris.core.persistence.models.ModelEntityChangeTracking</class>
-        <class>org.apache.polaris.core.persistence.models.ModelEntityDropped</class>
-        <class>org.apache.polaris.core.persistence.models.ModelGrantRecord</class>
-        <class>org.apache.polaris.core.persistence.models.ModelPrincipalSecrets</class>
-        <class>org.apache.polaris.core.persistence.models.ModelSequenceId</class>
+        <class>org.apache.polaris.jpa.models.ModelEntity</class>
+        <class>org.apache.polaris.jpa.models.ModelEntityActive</class>
+        <class>org.apache.polaris.jpa.models.ModelEntityChangeTracking</class>
+        <class>org.apache.polaris.jpa.models.ModelEntityDropped</class>
+        <class>org.apache.polaris.jpa.models.ModelGrantRecord</class>
+        <class>org.apache.polaris.jpa.models.ModelPrincipalSecrets</class>
+        <class>org.apache.polaris.jpa.models.ModelSequenceId</class>
         <shared-cache-mode>NONE</shared-cache-mode>
         <properties>
           <property name="jakarta.persistence.jdbc.url" value="jdbc:h2:mem:polaris-{realm}"/>
